scrapers/twitter/requirements.txt
requests-oauthlib~=1.3.1
pytz~=2021.3
python-dotenv~=0.20.0